added error handling
